+++ logprep/abc/connector.py
@@ -15,19 +15,19 @@ class Connector(Component):
 
         number_of_processed_events: CounterMetric = field(
             factory=lambda: CounterMetric(
-                description="Number of events that were processed",
+                description="Number of successfull events",
                 name="number_of_processed_events",
             )
         )
-        """Number of events that were processed"""
+        """Number of successfull events"""
 
         number_of_failed_events: CounterMetric = field(
             factory=lambda: CounterMetric(
-                description="Number of events that were send to error output",
+                description="Number of failed events",
                 name="number_of_failed_events",
             )
         )
-        """Number of events that were send to error output"""
+        """Number of failed events"""
 
         processing_time_per_event: HistogramMetric = field(
             factory=lambda: HistogramMetric(