+++ logprep/generator/http/input.py
@@ -9,6 +9,7 @@ import tempfile
 import time
 from datetime import datetime, timedelta
 from functools import cached_property
+from operator import itemgetter
 from pathlib import Path
 from typing import Generator, List
 
@@ -214,9 +215,7 @@ class Input:
             event_batch = sorted(event_batch, key=lambda x: x[0])
         log_classes = itertools.groupby(event_batch, key=lambda x: x[0])
         for target_path, events in log_classes:
-            request_data = "\n".join(self._encoder.encode(event).decode() for _, event in events)
-            event_target = f"{self.config.get('target_url')}{target_path}"
-            yield event_target, request_data
+            yield target_path, list(map(itemgetter(1), events))
 
     def _process_event_line(self, line):
         """