+++ logprep/registry.py
@@ -10,7 +10,8 @@ from logprep.connector.dummy.input import DummyInput
 from logprep.connector.dummy.output import DummyOutput
 from logprep.connector.elasticsearch.output import ElasticsearchOutput
 from logprep.connector.file.input import FileInput
-from logprep.connector.http.input import HttpConnector
+from logprep.connector.http.input import HttpInput
+from logprep.connector.http.output import HttpOutput
 from logprep.connector.json.input import JsonInput
 from logprep.connector.jsonl.input import JsonlInput
 from logprep.connector.jsonl.output import JsonlOutput
@@ -93,7 +94,8 @@ class Registry:
         "elasticsearch_output": ElasticsearchOutput,
         "jsonl_output": JsonlOutput,
         "opensearch_output": OpensearchOutput,
-        "http_input": HttpConnector,
+        "http_input": HttpInput,
+        "http_output": HttpOutput,
         "s3_output": S3Output,
     }