+++ tests/unit/test_run_logprep.py
@@ -326,7 +326,6 @@ class TestGeneratorCLI:
             events=None,
             shuffle=False,
             thread_count=1,
-            report=True,
             replace_timestamp=True,
             tag="loadtest",
             loglevel="INFO",
@@ -354,7 +353,7 @@ class TestGeneratorCLI:
                 "5000",
                 "--shuffle",
                 "False",
-                "--thread_count",
+                "--thread-count",
                 "2",
                 "--batch-size",
                 "1000",
@@ -364,11 +363,9 @@ class TestGeneratorCLI:
                 "test-tag",
                 "--loglevel",
                 "DEBUG",
-                "--report",
-                "False",
             ],
         )
-        assert result.exit_code == 0
+        assert result.exit_code == 0, result.stdout
         mock_generator.assert_called_with(
             input_dir="/some-path",
             target_url="some-domain",
@@ -378,7 +375,6 @@ class TestGeneratorCLI:
             events=5000,
             shuffle=False,
             thread_count=2,
-            report=False,
             replace_timestamp=False,
             tag="test-tag",
             loglevel="DEBUG",